+++ rose/__init__.py
@@ -1,4 +1,6 @@
 from rose.metrics import *
-from rose.learner import ActiveLearner
-from rose.data import InputFile, OutputFile
-from rose.engine import WorkflowEngine, ResourceEngine
+from rose.learner import Learner
+
+import rose.al.active_learner
+#import rose.rl.learner
+
